Add functional filters
With this patch, one is able to define a filter as a boolean function.
Thus a filter might be a string (so it works as before) or a function.

Just like before, the function is hooked to a single column.
The filtering function is given a single argument that is the value of a cell.

The function is applied to every row and the table retains only those which the filter returned a truthy value.

For now, the filters do not influence on each other, that is, they are applied to all the data and not only on those visible by the user, which might give some strange behavior at first glance.